+++ src/rules/embedTask.ts
@@ -1,0 +1,18 @@
+import customFence from "markdown-it-container";
+
+export default function task(md): void {
+  return customFence(md, "task", {
+    marker: "&",
+    validate: () => true,
+    render: function(tokens, idx) {
+      const { info } = tokens[idx];
+      if (tokens[idx].nesting === 1) {
+        // opening tag
+        return `<div class="task task-${md.utils.escapeHtml(info)}">\n`;
+      } else {
+        // closing tag
+        return "</div>\n";
+      }
+    },
+  });
+}